utils/config_loader.py
import yaml
from pathlib import Path


def load_config(config_path: str = "config/config.yaml") -> dict:
    """
    Load YAML configuration file.

    Parameters
    ----------
    config_path : str
        Path to the YAML config file.

    Returns
    -------
    dict
        Parsed configuration dictionary.
    """
    with open(Path(config_path), "r", encoding="utf-8") as f:
        config = yaml.safe_load(f)
    return config
