requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
joblib
nltk
wordcloud
scikit-plot
jupyter
ipykernel
pyyaml
python-dotenv
pyarrow
fastparquet
rank-bm25
